Fixed test table of contents toggle buttons
* Refactored the content page object to include new regions that
describe the page. The regions are Navbar, Sidebar, Sidebar header, and
Toolbar.
* Refactored the test to account for the differences in resolutions and
the order of toggling the Sidebar. When the resolution is desktop we
click the toc button on the sidebar to close and the toc button on the
toolbar to open. It's the other way around for mobile and tablet
resolutions.