hoist data attribue var setting to apply to both rex and cnx styles
If using dynamic styles, the .less rules for the css vars will not be
used, and we can't just set them in cnx-recipes because the cascade
for dynamic styles already starts at #main-content and we need to
check a data-attribute on that element. Moving them to the MainContent
component will ensure they are attached in both scenarios.