do not track highlight events if the card is not active
There is a bug in the existing analytics code that will cause double
"show create" and "show login" events when switching pages after a
legitimate event has fired. This is caused by the props.data being
a falsey value in the return check. This commit adds a second check
to make sure isActive is true before firing the event.