Fix some analytics issues (#1779)
* initial migration to gtag

* translate old params instead of changing all of them for now

also remove the unused trackEvent

* update specs and use eventAction as event name

* simplify set command

* fix type version

* use dimension directly due to issues with map not working

* update snapshot

* update prerender spec

* lint

* actually fix prerender test

* test more fixes

* fix test, initial page view should not have click event data

* tidy

* always turn off send_page_view for config calls

* temporarily add staging properties to dev env

* remove temp test ids

* ensure page title gets set before calling analytics

The body of the hook was extracted so it can be called directly
with await and it doesn't seem to actually need the action as
a param.

* organize imports

* fix lint

* use hook body directly, temporarily add staging test ids

* do not track highlight events if the card is not active

There is a bug in the existing analytics code that will cause double
"show create" and "show login" events when switching pages after a
legitimate event has fired. This is caused by the props.data being
a falsey value in the return check. This commit adds a second check
to make sure isActive is true before firing the event.

* remove test ids

Co-authored-by: staxly[bot] <35789409+staxly[bot]@users.noreply.github.com>